Fixed the vs generator linker dependencies.
The vs generator seemed to build a set of linker dependencies
but never added them to the 'Additional Dependencies' property.
This caused the generator to generate VS projects without
linker dependencies.

Removed the dependencies variable and added directly to the libs
variable since it isn't used for anything else anyway.