Add basic support for cross-compiling.
The target_platform and host_platform specifiers on generators have been
replaced with "target" and "host" System objects. These objects have a
"platform" property that corresponds to the v2.0 API "target_platform" and
"host_platform" properties. In addition, System objects have an "arch"
property that has a normalized CPU identifier.

Currently, all x86_64 CPU identifiers are normalized to "x86_64", and
all "x86" CPU identifiers are normalized to "x86".

Projects may specify the target architecture up-front by setting a
"target_arch" property on their BuildParser. If this property is not
specified, a "--target-arch" option will be available on the
command-line. These options both change the target.arch property on the
build.

Note that AMBuild makes no attempt to modify compiler flags based on the
target architecture. It is the responsibility of the user or configure/build
script to ensure that the compiler will be invoked correctly. However,
there is now a uniform way of communicating the host and desired
architecture to builders.

AMBuild does not yet attempt to check that the provided compiler
correctly produces output for the desired target. AMBuild also cannot yet
distinguish between host and target compilers. Finally, AMBuild does not
yet support changing the target platform. These features are left for
future work.