Overhaul the context design.
Context behavior is now limited based on intended use. There are four context types:
 - Empty, which cannot introduce outputs.
 - Build, which can introduce outputs.
 - TopLevel, which inherit from Build and can change their build folder.
 - Root, which inherit from Build and have no parent.

In addition odd edge cases that make the AMBuild output confusing have
been removed, such as creating separate context trees and abusing the
source folder hierarchy.

In terms of the front-facing API, there are now three entry-points to
creating new contexts:
 - Import(paths, vars) - Empty
 - Eval(paths, vars) - Empty
 - Build(paths, vars) - Build/TopLevel

In addition it is now possible to enter contexts outside of the
currentSourceFolder, if the parent context is the root or a TopLevel
context.