Fix a bug where discovered absolute paths were treated as source paths.
AMBuild canonicaly stores outputs as relative to the objdir, and inputs as absolute. If an external
dependency string contains an absolute path to the objdir, we must recognize this and fix the path,
lest we accidentally compute the new path as a source (which would leave us without a valid
ordering).