Simplify damage calculation.
A lot of this appears to be aggressively over optimizing. For example,
nodes were separated into "known dirty" and "maybe" dirty cases. Remove
this distinction, and mark everything dirty as long as it needs to be
marked. This also removes a hack in the Node object that was only used
for damage calculation.