Auto-detect Microsoft Visual C++.
This is a major refactoring of the C++ detection logic. The new behavior
is as follows:
 1. If 'CC' and 'CXX' are defined in the environment, those most result
    in a valid compiler, and no other detection is performed. On
    Windows, tools like "rc" and "lib" must be in the environment.
    Defining CC but not CXX (or vice-versa) is an error.
 2. If running on Windows, if 'cl.exe' is in PATH, then AMBuild assumes
    the environment has already been configured, and looks for, in this
    order: cl, clang, gcc, icc.
 3. If running on Windows, and 'cl.exe' is not in PATH, then AMBuild
    looks in the registry for Visual Studio 2015. It will also look for
    the 'vswhere' tool, and if present, will detect installations of
    Visual Studio 2017 and 2019. The highest working version is used. If
    no working version is found, the logic in step 2 is used as a
    last-ditch attempt.
 4. Otherwise, AMBuild tries to run clang, gcc, or icc (in that order).

This logic is heavily dependent on vcvars being functional. For example,
Visual Studio 2015 may not configure newer Windows Kits properly, and
AMBuild will not be able to automatically detect such an install.

The Preparer API has been bumped to 2.1.1. When this API version is
present, DetectCxx() now takes a dictionary of options for configuring
C++ detection. This will be documented in detail on the wiki (currently
it has only one option). To allow backwards compatibility, projects can
use the following snippet to detect 2.1.1:

    apiVersion = getattr(builder, 'apiVerison', None)
    if apiVersion and apiVersion >= '2.1.1':
    	cxx = builder.DetectCxx(cxx_options)
    else:
    	cxx = builder.DetectCxx()

Some important changes have been made underneath the hood. vcvars
configures a large number of extra environment variables. Normally we
would cache these values globally, and AMBuild would restore the
original environment on each build. The existing mechanism to do this is
too limited. Additionally, having a global environment means AMBuild can
only ever support one compiler per build. Windows has compiler binaries
for X64, ARM, ARM64, and X86 support, so universal builds would be
impossible if we kept this model.

We do not yet support universal builds in AMBuild. However, this patch
takes an important first step. Compiler objects can now have an
"environment" attached, which contains two critical pieces of
information:
 1. An environment command list. This is a sequence of edit operations
    to the environment, which we derive by comparing the environment
    before and after calling vcvars.
 2. A dictionary mapping tool names to their executable paths. This way,
    we can attach the exact paths for cl.exe, rc.exe, and lib.exe.
    Knowing the full path to a tool means we don't have to touch the
    global environment of each worker process. It also (slightly) makes
    the build process more hermetic, and significantly reduces the database
    size by not attaching full path information to thousands of C++ commands.

Environment data is stored in the database, and changes to an
environment will trigger a rebuild of any affected commands.

This is an important step forward because it will enable a number of
additional features to AMBuild. Most importantly, it enables future
support for multiple compilers/targets. It also opens the door for more
hermetic builds: we can re-build the environment from the database
during automatic reconfigures.

Finally, some quality improvements have been made. The logic for
flagging nodes as dirty has been greatly simplified, and should be more
accurate as a result.