Introduce the AMBuild 2.2 API.
The major feature with this API is the ability to run multiple C++
compilers. This is important for supporting mixed builds in one objdir,
as it allows us to build unified packages.

We can already do this on Linux/Mac by adding -m32/-m64 to the
command-line as needed. But this doesn't work on Windows, where we have
to actually invoke separate versions of cl.exe.

Unfortunately, this feature comes at a cost, which is that the API has
to change significantly. The concept of a per-builder target no longer
makes sense. The target has to be tied to the compiler. This also means
we can no longer tie compilers to builders, which was convenient but
never made much sense to begin with.

Retaining some amount of convenience is important as we don't want
AMBuild scripts to become any more boilerplate heavy than they already
are. Boilerplate is fine for huge projects like SourceMod which have
very complicated builds. But for small projects, the more declarative
the API, the better.

Thus, this API is still very much a work-in-progress, and will remain so
until we can figure out how to conveniently communicate compiler objects
to build scripts.