Refactor the Projects API. (#101)
The Projects API was designed for multi-architecture use, back when we
had to hack support for that in each build script. Now, they need to be
decoupled from compilers to properly support multi-arch builds.

This patch removes the interface for projects off the C++ compiler and
into the build context, meaning, it will go through the generator. This
will also make vsgen support of the 2.2 API a little cleaner.