Add a way to write files as part of the build.
Sometimes it's useful to create a one-off file as part of the build,
such as an auto-generated header or something as part of a packaging
step.

Previously, the only way to do this was to add a bunch of boilerplate,
creating a command node around something like "echo", or by having a
custom script. The boilerplate around AddCommand() is annoying, and it's
tempting to simply open() and write a file, but this would not properly
integrate with the dependency graph.

The new API, "AddOutputFile", allows build scripts to specify a file
name and its binary contents. This creates a job in the build and
returns a node that can be used as a dependency, without any
boilerplate. (Other than, it is necessary to explicitly encode any
string contents).