Add support for precompiled headers.
A new builder type, PrecompiledHeaders, has been added. It is virtually
identical to a Program, Library, or StaticLibrary builder, except the
list of source files should be a list of header files. Once added to the
build context, it produces a PchNode, which can be added to a list of
includes on another object.

Windows developers will be familiar with the Microsoft approach, which
is to create a "stdafx.h" file including everything to be precompiled.
Clang and GCC effectively require the same kind of thing. AMBuild
abstracts this process for you, by creating a "unified header". For
example:

    pch_builder = cxx.PrecompiledHeaders('all-headers', 'c++')
    pch_builder.sources += ['vector', 'string', 'string.h', 'iostream']
    pch = builder.Add(pch_builder)

    binary = cxx.Library('module1')
    binary.cxxincludes += [pch]

This will create a unified 'all-headers.h' file, equivalent to a
stdafx.h file in Visual Studio, which #includes all of the header files
under `pch_builder.sources`. To take advantage of the precompiled
header, each source file should include it as the very first directive:

    #include <all-headers.h>

Merely including a sub-header, like string.h, will not use the
precompiled header.

It is currently not possible to use more than one precompiled header
object per module. AMBuild will let you do this, but the compiler will
likely ignore anything but the first PCH on each translation unit.

Although MSVC, GCC, and Clang all support some notion of precompiled
headers, it is a very esoteric feature and comes with many
vendor-specific warnings. Caveat emptor.

Implementation details:

If there are no C++ tokens and an #include directive is reached, GCC will
look for a .gch file in include order.

Clang requires .pch files to be manually injected on the command-line.

MSVC behaves differently with /Z[iI] (pdb files) vs /Z7. With PDB files,
the PDB from the PCH must be copied into the folder being used to create
object files. This is very complex to handle properly so instead we rely
on /Z7. In this mode, we can create a .cpp that includes the unified
header, compile it, and link the object file into the final DLL/EXE that
relies on a PCH.