Switch from /Zi to /Z7.
/Zi is incompatible with the way AMBuild uses MSVC, since rather than
issue one cl.exe invocation with all source files, we issue separate
invocations in parallel. The documentation suggests that /FS is not
reliable enough as a workaround. It's not clear why it exists then, but
it's best that we use an alternative, /Z7.

Consumers of the 2.2 API will now see /Z7 used by default. This means
object files will have embedded PDB info, and "vc140.pdb" files are no
longer emitted.

If for some reason this turns out to be disadvantageous, we can look
into per-object pdb files, or switching how cl.exe is invoked.