Switch gcc/clang to use -MD for dependency discovery. (#123)
Currently we use -H for gcc/clang dependency discovery, which has a few
problems. The output format is not well defined, it omits command-line
headers, and is inconsistent with how precompiled headers are treated.

Using -MD means we have to track an additional output from every gcc or
clang job. Cxx tasks now have a "deps" field in the database which can
store extended information on how to parse dependencies. For this case,
it stores a tuple ('md', md_file).

While we're here, ObjectFile has been renamed to ObjectFileTask. The
purpose of this class has always been extremely confusing and this
clarifies it a bit.

Finally, AMBuild now has a license file. Originally I was going to
import pymake and at the time needed an actual license to do this, but
pymake turned out to be difficult to embed. So now we have a very
limited -MD parser, and a license.

Bug: fixes issue #65