Use arm64 MSVC build tools even when using the x64 version of Python on the arm64 version of Windows 11 and fallback to x64 or x86 versions if not available.
On the arm64 version of Windows 11, platform.machine() returns 'AMD64' when using the x64 version of
python. To work around this, IsWow64Process2 is used to determine the native architecture.

Fallback to the x64 or x86 build tools is also enabled if the arm64 version of the tools is not
installed. Build tool candidates are only considered if the architecture of the binaries is
executable on the current system. This is notable because x64 binaries cannot be run on the arm64
version of Windows 10 but can be run on Windows 11. The arm version of Windows 10 only has emulation
support for x86 binaries. The GetMachineTypeAttributes and IsWow64GuestMachineSupported APIs are used
to deal with these possibilties.