Add a Protoc C++ tool.
Example:

    prog = cxx.Program('prog')
    protos = builder.tools.Protoc(sources = ['myproto.proto'])
    prog.custom += [protos]

To make this work, protoc is now always executed in the top level of the
build directory. This is kind of gross as we have to make sure all paths
are relative to the root, rather than cwd. The alternative is running
protoc once for each language, which would be fine, but the current
thing is working.